payroll/requirements.txt
amqp==2.5.2
apturl==0.5.2
asgiref==3.2.5
asn1crypto==0.24.0
billiard==3.6.3.0
Brlapi==0.6.6
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
Django==3.0.3
django-ipware==2.1.0
django-picklefield==2.1.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
get==2019.4.13
html5lib==1.0.1
httplib2==0.9.2
idna==2.9
importlib-metadata==1.6.0
keyring==10.6.0
keyrings.alt==3.0
kombu==4.6.8
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==7.0.0
post==2019.4.13
protobuf==3.0.0
public==2019.4.13
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
PyPDF2==1.26.0
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.6.1
python-debian==0.1.32
python-http-client==3.2.6
pytz==2019.3
pyxdg==0.25
PyYAML==3.12
query-string==2019.4.13
redis==3.4.1
reportlab==3.5.34
request==2019.4.13
requests==2.23.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
sendgrid==6.2.0
simplejson==3.13.2
six==1.14.0
sqlparse==0.3.1
system-service==0.3
systemd-python==234
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
vine==1.3.0
wadllib==1.3.2
webencodings==0.5.1
xhtml2pdf==0.2.4
xkit==0.0.0
zipp==3.1.0
zope.interface==4.3.2
